feat: Many functions are created

+++ backend/src/controllers/notification.controller.js
@@ -1,22 +1,34 @@
 import Notification from "../models/Notification.js";
 
-export const getNotifications = async (req, res) => {
+export const getMyNotifications = async (req, res) => {
 
     try {
 
         const notifications = await Notification.find({
 
-            receiver: req.user._id
+            user: req.user._id
 
-        }).sort({
+        })
+
+        .sort({
 
             createdAt: -1
+
+        });
+
+        const unreadCount = await Notification.countDocuments({
+
+            user: req.user._id,
+
+            isRead: false
 
         });
 
         return res.status(200).json({
 
             success: true,
+
+            unreadCount,
 
             notifications
 
@@ -60,6 +72,22 @@ export const markAsRead = async (req, res) => {
 
         }
 
+        if (
+
+            notification.user.toString() !== req.user._id.toString()
+
+        ) {
+
+            return res.status(403).json({
+
+                success: false,
+
+                message: "Access denied"
+
+            });
+
+        }
+
         notification.isRead = true;
 
         await notification.save();
@@ -68,7 +96,9 @@ export const markAsRead = async (req, res) => {
 
             success: true,
 
-            message: "Notification marked as read"
+            message: "Notification marked as read",
+
+            notification
 
         });
 
@@ -87,6 +117,58 @@ export const markAsRead = async (req, res) => {
     }
 
 };
+
+
+export const markAllAsRead = async (req, res) => {
+
+    try {
+
+        await Notification.updateMany(
+
+            {
+
+                user: req.user._id,
+
+                isRead: false
+
+            },
+
+            {
+
+                $set: {
+
+                    isRead: true
+
+                }
+
+            }
+
+        );
+
+        return res.status(200).json({
+
+            success: true,
+
+            message: "All notifications marked as read"
+
+        });
+
+    } catch (error) {
+
+        console.log(error);
+
+        return res.status(500).json({
+
+            success: false,
+
+            message: "Internal Server Error"
+
+        });
+
+    }
+
+};
+
 
 export const deleteNotification = async (req, res) => {
 
@@ -110,6 +192,22 @@ export const deleteNotification = async (req, res) => {
 
         }
 
+        if (
+
+            notification.user.toString() !== req.user._id.toString()
+
+        ) {
+
+            return res.status(403).json({
+
+                success: false,
+
+                message: "Access denied"
+
+            });
+
+        }
+
         await Notification.findByIdAndDelete(
 
             req.params.id
@@ -120,7 +218,7 @@ export const deleteNotification = async (req, res) => {
 
             success: true,
 
-            message: "Notification deleted"
+            message: "Notification deleted successfully"
 
         });
